Allow setting container cpu and memory (#18)
Parameterize container CPU and memory to allow users to right size containers.  The values need to be a matched pair[1]

When using fargate the memory and cpu are required at the ECS task level instead of at the container level[2].  Refactored to set memory and cpu at the task level and removed them from container level.

[1] https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task_definition_parameters.html
[2] https://docs.aws.amazon.com/AWSCloudFormation/latest/TemplateReference/aws-resource-ecs-taskdefinition.html